Depth data is added in dataset + getitem + PC projection is calcualted as pseudo mask

+++ custom_models/helpers/load_depth_image.py
@@ -1,4 +1,5 @@
 from PIL import Image
+import torch
 from torchvision.transforms import ToTensor
 
 def load_depth_image(video_frames, size_x_y):
@@ -14,6 +15,6 @@ def load_depth_image(video_frames, size_x_y):
         resize_shape = [size_x_y[0], gt_depth_image.size[1] // scaling_factor]  # It is PIL reshape give it as W,H
         # Resize depth image to the specified size
         gt_depth_image = gt_depth_image.resize(resize_shape)
-        gt_depth_image = to_tensor(gt_depth_image)
+        gt_depth_image = to_tensor(gt_depth_image).type(torch.float32)  # Convert to tensor and float32
         depth_images.append(gt_depth_image)
     return depth_images
